pyproject.toml
[tool.poetry]
name = "mr-backend"
version = "0.1.0"
description = ""
authors = ["MaHryCT3 <mahryct123@gmail.com>"]
readme = "README.md"
packages = [{include = "mr_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.94.1"
httpx = "^0.23.3"
pydantic = "^1.10.6"
sqlalchemy = "^2.0.6"
asyncpg = "^0.27.0"
loguru = "^0.6.0"
uvicorn = {extras = ["standard"], version = "^0.21.0"}
alembic = "^1.10.2"
pyjwt = "^2.6.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.256"
black = "^23.1.0"
mypy = "^1.1.1"
pytest = "^7.2.2"
pytest-aio = "^1.5.0"
pytest-dotenv = "^0.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
select = ['Q', 'F', 'I', 'ANN', 'B', 'C4', 'TID', 'ARG', 'PLR']
ignore = ['ANN101', 'ANN204', 'ANN401', 'ANN002', 'ANN003', 'PLR0913']  
# ANN - https://beta.ruff.rs/docs/rules/#flake8-annotations-ann, 
# PLR0913 - Too many arguments


[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"] # Unused import
"tests/*" = ['ANN']  # Annotations
"app/tools/time.py" = ['PLR2004']
"migration/*" = ['F']
"app/api/*" = ['B008', 'ARG001']

[tool.ruff.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "single"

[tool.black]
skip-string-normalization = true
exclude = '(\.git|\.venv|\.vscode)'
line-length = 120

[tool.mypy]
python_version = '3.11'
ignore_missing_imports = true
warn_no_return = false
warn_return_any = false
warn_unused_configs = true
disable_error_code = 'assignment, union-attr, override'

